tutorials/Learning Series 11/requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
matplotlib>=3.8
pyyaml>=6.0
